+++ client/src/components/Common/Button/type.ts
@@ -1,4 +1,7 @@
+
+
 export interface ButtonType {
+  id?: number;
   value?: string;
   type?: string;
   width?: string;
