--- client/src/components/Common/HashBtn/type.ts
@@ -1,4 +1,0 @@
-export interface HashType {
-  value?: string;
-  onClick?: React.MouseEventHandler<HTMLButtonElement> | void;
-}
