--- client/src/pages/Detail/Writecontent.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Writecontent = () => {
-  return <div>Writecontent</div>;
-};
-
-export default Writecontent;
