--- client/src/pages/Writecontent/type.ts
@@ -1,4 +1,0 @@
-export interface WriteContentType {
-	onClickImg?: React.MouseEventHandler<HTMLImageElement> | void;
-  }
-
